[#3] Refactor typeclasses to match type-natural
Problem: The definitions and the instance declarations of the IsCommutativeRing
and IsInteger typeclasses are misplaced with respect to the structure in
type-natural, making interoperability more difficult down the line.

Solution: Rearrange the definitions and instance declarations to match
those of IsPeano and PeanoOrder in type-natural.
